requirements.txt
gspread>=6.0.0
google-auth>=2.23.0
eventregistry>=9.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pytz>=2023.3

